Made samples self-contained
