+++ sdks/python/apache_beam/examples/complete/game/game_stats.py
@@ -18,7 +18,7 @@
 """Fourth in a series of four pipelines that tell a story in a 'gaming' domain.
 
 --------------------------------------------------------------------------------
-NOTE: This example is not yet runnable by Dataflow. The runner still needs
+NOTE: This example is not yet runnable by DataflowRunner. The runner still needs
       support for:
         * the --save_main_session flag when streaming is enabled
         * combiners
@@ -58,8 +58,6 @@ To specify a different runner:
 
 NOTE: When specifying a different runner, additional runner-specific options
       may have to be passed in as well
-NOTE: With DataflowRunner, the --setup_file flag must be specified to handle the
-      'util' module
 
 EXAMPLES
 --------
@@ -76,7 +74,6 @@ python game_stats.py \
     --topic projects/$PROJECT_ID/topics/$PUBSUB_TOPIC \
     --dataset $BIGQUERY_DATASET \
     --runner DataflowRunner \
-    --setup_file ./setup.py \
     --staging_location gs://$BUCKET/user_score/staging \
     --temp_location gs://$BUCKET/user_score/temp
 """
@@ -84,16 +81,118 @@ python game_stats.py \
 from __future__ import absolute_import
 
 import argparse
+import csv
 import logging
+import time
 
 import apache_beam as beam
-from apache_beam.transforms import combiners
-
-from apache_beam.examples.complete.game.util import util
+from apache_beam.metrics.metric import Metrics
 from apache_beam.options.pipeline_options import GoogleCloudOptions
 from apache_beam.options.pipeline_options import PipelineOptions
 from apache_beam.options.pipeline_options import SetupOptions
 from apache_beam.options.pipeline_options import StandardOptions
+
+
+class ParseGameEventFn(beam.DoFn):
+  """Parses the raw game event info into a Python dictionary.
+
+  Each event line has the following format:
+    username,teamname,score,timestamp_in_ms,readable_time
+
+  e.g.:
+    user2_AsparagusPig,AsparagusPig,10,1445230923951,2015-11-02 09:09:28.224
+
+  The human-readable time string is not used here.
+  """
+  def __init__(self):
+    super(ParseGameEventFn, self).__init__()
+    self.num_parse_errors = Metrics.counter(self.__class__, 'num_parse_errors')
+
+  def process(self, elem):
+    try:
+      row = list(csv.reader([elem]))[0]
+      yield {
+          'user': row[0],
+          'team': row[1],
+          'score': int(row[2]),
+          'timestamp': int(row[3]) / 1000.0,
+      }
+    except:  # pylint: disable=bare-except
+      # Log and count parse errors
+      self.num_parse_errors.inc()
+      logging.error('Parse error on "%s"', elem)
+
+
+class ExtractAndSumScore(beam.PTransform):
+  """A transform to extract key/score information and sum the scores.
+  The constructor argument `field` determines whether 'team' or 'user' info is
+  extracted.
+  """
+  def __init__(self, field):
+    super(ExtractAndSumScore, self).__init__()
+    self.field = field
+
+  def expand(self, pcoll):
+    return (pcoll
+            | beam.Map(lambda elem: (elem[self.field], elem['score']))
+            | beam.CombinePerKey(sum))
+
+
+class TeamScoresDict(beam.DoFn):
+  """Formats the data into a dictionary of BigQuery columns with their values
+
+  Receives a (team, score) pair, extracts the window start timestamp, and
+  formats everything together into a dictionary. The dictionary is in the format
+  {'bigquery_column': value}
+  """
+  def process(self, team_score, window=beam.DoFn.WindowParam):
+    team, score = team_score
+    start = timestamp2str(int(window.start))
+    yield {
+        'team': team,
+        'total_score': score,
+        'window_start': start,
+        'processing_time': timestamp2str(int(time.time()))
+    }
+
+
+class WriteToBigQuery(beam.PTransform):
+  """Generate, format, and write BigQuery table row information."""
+  def __init__(self, table_name, dataset, schema):
+    """Initializes the transform.
+    Args:
+      table_name: Name of the BigQuery table to use.
+      dataset: Name of the dataset to use.
+      schema: Dictionary in the format {'column_name': 'bigquery_type'}
+    """
+    super(WriteToBigQuery, self).__init__()
+    self.table_name = table_name
+    self.dataset = dataset
+    self.schema = schema
+    self.create_disposition = beam.io.BigQueryDisposition.CREATE_IF_NEEDED
+    self.write_disposition = beam.io.BigQueryDisposition.WRITE_APPEND
+
+  def get_bigquery_schema(self):
+    """Build the output table schema."""
+    return ', '.join(
+        '%s:%s' % (col, self.schema[col]) for col in self.schema)
+
+  def get_table(self, pipeline):
+    """Utility to construct an output table reference."""
+    project = pipeline.options.view_as(GoogleCloudOptions).project
+    return '%s:%s.%s' % (project, self.dataset, self.table_name)
+
+  def expand(self, pcoll):
+    table = self.get_table(pcoll.pipeline)
+    return (
+        pcoll
+        | 'ConvertToRow' >> beam.Map(
+            lambda elem: {col: elem[col] for col in self.schema})
+        | beam.io.Write(beam.io.BigQuerySink(
+            table,
+            schema=self.get_bigquery_schema(),
+            create_disposition=self.create_disposition,
+            write_disposition=self.write_disposition)))
 
 
 class CalculateSpammyUsers(beam.PTransform):
@@ -118,7 +217,7 @@ class CalculateSpammyUsers(beam.PTransform):
         sum_scores
         | beam.Values()
         # NOTE: DataflowRunner currently does not support combiners
-        | beam.CombineGlobally(combiners.MeanCombineFn()).as_singleton_view()
+        | beam.Mean.globally().as_singleton_view()
     )
 
     # Filter the user sums using the global mean.
@@ -199,7 +298,7 @@ def run(argv=None):
     raw_events = (
         p
         | 'ReadPubSub' >> beam.io.gcp.pubsub.ReadStringsFromPubSub(args.topic)
-        | 'ParseGameEventFn' >> beam.ParDo(util.ParseGameEventFn())
+        | 'ParseGameEventFn' >> beam.ParDo(ParseGameEventFn())
         | 'AddEventTimestamps' >> beam.Map(
             lambda elem: beam.window.TimestampedValue(elem, elem['timestamp']))
     )
@@ -225,8 +324,7 @@ def run(argv=None):
         # Derive a view from the collection of spammer users. It will be used as
         # a side input in calculating the team score sums, below
         # NOTE: DataflowRunner currently does not support combiners
-        | 'CreateSpammersView' >> beam.CombineGlobally(
-            combiners.ToDictCombineFn()).as_singleton_view()
+        | 'CreateSpammersView' >> beam.ToDict.globally().as_singleton_view()
     )
 
     # Calculate the total score per team over fixed windows, and emit cumulative
@@ -248,10 +346,10 @@ def run(argv=None):
          lambda elem, spammers: elem['user'] not in spammers,
          spammers_view)
      # Extract and sum teamname/score pairs from the event data.
-     | 'ExtractAndSumScore' >> util.ExtractAndSumScore('team')
-     | 'TeamScoresDict' >> beam.ParDo(util.TeamScoresDict())
+     | 'ExtractAndSumScore' >> ExtractAndSumScore('team')
+     | 'TeamScoresDict' >> beam.ParDo(TeamScoresDict())
      # Write the result to BigQuery
-     | 'WriteTeamScoreSums' >> util.WriteToBigQuery(
+     | 'WriteTeamScoreSums' >> WriteToBigQuery(
          args.table_name + '_teams', args.dataset, teams_schema)
     )
 
@@ -282,10 +380,10 @@ def run(argv=None):
 
      # Find the mean session duration in each window
      # NOTE: DataflowRunner currently does not support combiners
-     | beam.CombineGlobally(combiners.MeanCombineFn()).without_defaults()
+     | beam.Mean.globally().without_defaults()
      | 'FormatAvgSessionLength' >> beam.Map(
          lambda elem: {'mean_duration': float(elem)})
-     | 'WriteAvgSessionLength' >> util.WriteToBigQuery(
+     | 'WriteAvgSessionLength' >> WriteToBigQuery(
          args.table_name + '_sessions', args.dataset, sessions_schema)
     )
