+++ sdks/python/apache_beam/examples/complete/game/hourly_team_score.py
@@ -47,8 +47,6 @@ To specify a different runner:
 
 NOTE: When specifying a different runner, additional runner-specific options
       may have to be passed in as well
-NOTE: With DataflowRunner, the --setup_file flag must be specified to handle the
-      'util' module
 
 EXAMPLES
 --------
@@ -63,7 +61,6 @@ python hourly_team_score.py \
     --project $PROJECT_ID \
     --dataset $BIGQUERY_DATASET \
     --runner DataflowRunner \
-    --setup_file ./setup.py \
     --staging_location gs://$BUCKET/user_score/staging \
     --temp_location gs://$BUCKET/user_score/temp
 """
@@ -71,27 +68,81 @@ python hourly_team_score.py \
 from __future__ import absolute_import
 
 import argparse
+import csv
 import logging
+import time
+from datetime import datetime
 
 import apache_beam as beam
-
-from apache_beam.examples.complete.game.util import util
+from apache_beam.metrics.metric import Metrics
 from apache_beam.options.pipeline_options import GoogleCloudOptions
 from apache_beam.options.pipeline_options import PipelineOptions
 from apache_beam.options.pipeline_options import SetupOptions
 
 
+def str2timestamp(s, fmt='%Y-%m-%d-%H-%M'):
+  """Converts a string into a unix timestamp."""
+  dt = datetime.strptime(s, fmt)
+  epoch = datetime.utcfromtimestamp(0)
+  return (dt - epoch).total_seconds()
+
+
+class ParseGameEventFn(beam.DoFn):
+  """Parses the raw game event info into a Python dictionary.
+
+  Each event line has the following format:
+    username,teamname,score,timestamp_in_ms,readable_time
+
+  e.g.:
+    user2_AsparagusPig,AsparagusPig,10,1445230923951,2015-11-02 09:09:28.224
+
+  The human-readable time string is not used here.
+  """
+  def __init__(self):
+    super(ParseGameEventFn, self).__init__()
+    self.num_parse_errors = Metrics.counter(self.__class__, 'num_parse_errors')
+
+  def process(self, elem):
+    try:
+      row = list(csv.reader([elem]))[0]
+      yield {
+          'user': row[0],
+          'team': row[1],
+          'score': int(row[2]),
+          'timestamp': int(row[3]) / 1000.0,
+      }
+    except:  # pylint: disable=bare-except
+      # Log and count parse errors
+      self.num_parse_errors.inc()
+      logging.error('Parse error on "%s"', elem)
+
+
+class ExtractAndSumScore(beam.PTransform):
+  """A transform to extract key/score information and sum the scores.
+  The constructor argument `field` determines whether 'team' or 'user' info is
+  extracted.
+  """
+  def __init__(self, field):
+    super(ExtractAndSumScore, self).__init__()
+    self.field = field
+
+  def expand(self, pcoll):
+    return (pcoll
+            | beam.Map(lambda elem: (elem[self.field], elem['score']))
+            | beam.CombinePerKey(sum))
+
+
 class HourlyTeamScore(beam.PTransform):
   def __init__(self, start_min, stop_min, window_duration):
     super(HourlyTeamScore, self).__init__()
-    self.start_timestamp = util.str2timestamp(start_min)
-    self.stop_timestamp = util.str2timestamp(stop_min)
+    self.start_timestamp = str2timestamp(start_min)
+    self.stop_timestamp = str2timestamp(stop_min)
     self.window_duration_in_seconds = window_duration * 60
 
   def expand(self, pcoll):
     return (
         pcoll
-        | 'ParseGameEventFn' >> beam.ParDo(util.ParseGameEventFn())
+        | 'ParseGameEventFn' >> beam.ParDo(ParseGameEventFn())
 
         # Filter out data before and after the given times so that it is not
         # included in the calculations. As we collect data in batches (say, by
@@ -114,8 +165,65 @@ class HourlyTeamScore(beam.PTransform):
             beam.window.FixedWindows(self.window_duration_in_seconds))
 
         # Extract and sum teamname/score pairs from the event data.
-        | 'ExtractAndSumScore' >> util.ExtractAndSumScore('team')
+        | 'ExtractAndSumScore' >> ExtractAndSumScore('team')
     )
+
+
+class TeamScoresDict(beam.DoFn):
+  """Formats the data into a dictionary of BigQuery columns with their values
+
+  Receives a (team, score) pair, extracts the window start timestamp, and
+  formats everything together into a dictionary. The dictionary is in the format
+  {'bigquery_column': value}
+  """
+  def process(self, team_score, window=beam.DoFn.WindowParam):
+    team, score = team_score
+    start = timestamp2str(int(window.start))
+    yield {
+        'team': team,
+        'total_score': score,
+        'window_start': start,
+        'processing_time': timestamp2str(int(time.time()))
+    }
+
+
+class WriteToBigQuery(beam.PTransform):
+  """Generate, format, and write BigQuery table row information."""
+  def __init__(self, table_name, dataset, schema):
+    """Initializes the transform.
+    Args:
+      table_name: Name of the BigQuery table to use.
+      dataset: Name of the dataset to use.
+      schema: Dictionary in the format {'column_name': 'bigquery_type'}
+    """
+    super(WriteToBigQuery, self).__init__()
+    self.table_name = table_name
+    self.dataset = dataset
+    self.schema = schema
+    self.create_disposition = beam.io.BigQueryDisposition.CREATE_IF_NEEDED
+    self.write_disposition = beam.io.BigQueryDisposition.WRITE_APPEND
+
+  def get_bigquery_schema(self):
+    """Build the output table schema."""
+    return ', '.join(
+        '%s:%s' % (col, self.schema[col]) for col in self.schema)
+
+  def get_table(self, pipeline):
+    """Utility to construct an output table reference."""
+    project = pipeline.options.view_as(GoogleCloudOptions).project
+    return '%s:%s.%s' % (project, self.dataset, self.table_name)
+
+  def expand(self, pcoll):
+    table = self.get_table(pcoll.pipeline)
+    return (
+        pcoll
+        | 'ConvertToRow' >> beam.Map(
+            lambda elem: {col: elem[col] for col in self.schema})
+        | beam.io.Write(beam.io.BigQuerySink(
+            table,
+            schema=self.get_bigquery_schema(),
+            create_disposition=self.create_disposition,
+            write_disposition=self.write_disposition)))
 
 
 def run(argv=None):
@@ -182,8 +290,8 @@ def run(argv=None):
      | 'ReadInputText' >> beam.io.ReadFromText(args.input)
      | 'HourlyTeamScore' >> HourlyTeamScore(
          args.start_min, args.stop_min, args.window_duration)
-     | 'TeamScoresDict' >> beam.ParDo(util.TeamScoresDict())
-     | 'WriteTeamScoreSums' >> util.WriteToBigQuery(
+     | 'TeamScoresDict' >> beam.ParDo(TeamScoresDict())
+     | 'WriteTeamScoreSums' >> WriteToBigQuery(
          args.table_name, args.dataset, schema)
     )
