+++ sdks/python/apache_beam/examples/complete/game/leader_board.py
@@ -18,7 +18,7 @@
 """Third in a series of four pipelines that tell a story in a 'gaming' domain.
 
 --------------------------------------------------------------------------------
-NOTE: This example is not yet runnable by Dataflow. The runner still needs
+NOTE: This example is not yet runnable by DataflowRunner. The runner still needs
       support for:
         * the --save_main_session flag when streaming is enabled
 --------------------------------------------------------------------------------
@@ -65,8 +65,6 @@ To specify a different runner:
 
 NOTE: When specifying a different runner, additional runner-specific options
       may have to be passed in as well
-NOTE: With DataflowRunner, the --setup_file flag must be specified to handle the
-      'util' module
 
 EXAMPLES
 --------
@@ -83,7 +81,6 @@ python leader_board.py \
     --topic projects/$PROJECT_ID/topics/$PUBSUB_TOPIC \
     --dataset $BIGQUERY_DATASET \
     --runner DataflowRunner \
-    --setup_file ./setup.py \
     --staging_location gs://$BUCKET/user_score/staging \
     --temp_location gs://$BUCKET/user_score/temp
 """
@@ -91,16 +88,119 @@ python leader_board.py \
 from __future__ import absolute_import
 
 import argparse
+import csv
 import logging
+import time
 
 import apache_beam as beam
-from apache_beam.transforms import trigger
-
-from apache_beam.examples.complete.game.util import util
+from apache_beam.metrics.metric import Metrics
 from apache_beam.options.pipeline_options import GoogleCloudOptions
 from apache_beam.options.pipeline_options import PipelineOptions
 from apache_beam.options.pipeline_options import SetupOptions
 from apache_beam.options.pipeline_options import StandardOptions
+from apache_beam.transforms import trigger
+
+
+class ParseGameEventFn(beam.DoFn):
+  """Parses the raw game event info into a Python dictionary.
+
+  Each event line has the following format:
+    username,teamname,score,timestamp_in_ms,readable_time
+
+  e.g.:
+    user2_AsparagusPig,AsparagusPig,10,1445230923951,2015-11-02 09:09:28.224
+
+  The human-readable time string is not used here.
+  """
+  def __init__(self):
+    super(ParseGameEventFn, self).__init__()
+    self.num_parse_errors = Metrics.counter(self.__class__, 'num_parse_errors')
+
+  def process(self, elem):
+    try:
+      row = list(csv.reader([elem]))[0]
+      yield {
+          'user': row[0],
+          'team': row[1],
+          'score': int(row[2]),
+          'timestamp': int(row[3]) / 1000.0,
+      }
+    except:  # pylint: disable=bare-except
+      # Log and count parse errors
+      self.num_parse_errors.inc()
+      logging.error('Parse error on "%s"', elem)
+
+
+class ExtractAndSumScore(beam.PTransform):
+  """A transform to extract key/score information and sum the scores.
+  The constructor argument `field` determines whether 'team' or 'user' info is
+  extracted.
+  """
+  def __init__(self, field):
+    super(ExtractAndSumScore, self).__init__()
+    self.field = field
+
+  def expand(self, pcoll):
+    return (pcoll
+            | beam.Map(lambda elem: (elem[self.field], elem['score']))
+            | beam.CombinePerKey(sum))
+
+
+class TeamScoresDict(beam.DoFn):
+  """Formats the data into a dictionary of BigQuery columns with their values
+
+  Receives a (team, score) pair, extracts the window start timestamp, and
+  formats everything together into a dictionary. The dictionary is in the format
+  {'bigquery_column': value}
+  """
+  def process(self, team_score, window=beam.DoFn.WindowParam):
+    team, score = team_score
+    start = timestamp2str(int(window.start))
+    yield {
+        'team': team,
+        'total_score': score,
+        'window_start': start,
+        'processing_time': timestamp2str(int(time.time()))
+    }
+
+
+class WriteToBigQuery(beam.PTransform):
+  """Generate, format, and write BigQuery table row information."""
+  def __init__(self, table_name, dataset, schema):
+    """Initializes the transform.
+    Args:
+      table_name: Name of the BigQuery table to use.
+      dataset: Name of the dataset to use.
+      schema: Dictionary in the format {'column_name': 'bigquery_type'}
+    """
+    super(WriteToBigQuery, self).__init__()
+    self.table_name = table_name
+    self.dataset = dataset
+    self.schema = schema
+    self.create_disposition = beam.io.BigQueryDisposition.CREATE_IF_NEEDED
+    self.write_disposition = beam.io.BigQueryDisposition.WRITE_APPEND
+
+  def get_bigquery_schema(self):
+    """Build the output table schema."""
+    return ', '.join(
+        '%s:%s' % (col, self.schema[col]) for col in self.schema)
+
+  def get_table(self, pipeline):
+    """Utility to construct an output table reference."""
+    project = pipeline.options.view_as(GoogleCloudOptions).project
+    return '%s:%s.%s' % (project, self.dataset, self.table_name)
+
+  def expand(self, pcoll):
+    table = self.get_table(pcoll.pipeline)
+    return (
+        pcoll
+        | 'ConvertToRow' >> beam.Map(
+            lambda elem: {col: elem[col] for col in self.schema})
+        | beam.io.Write(beam.io.BigQuerySink(
+            table,
+            schema=self.get_bigquery_schema(),
+            create_disposition=self.create_disposition,
+            write_disposition=self.write_disposition)))
 
 
 class CalculateTeamScores(beam.PTransform):
@@ -128,7 +228,7 @@ class CalculateTeamScores(beam.PTransform):
                                            trigger.AfterCount(20)),
             accumulation_mode=trigger.AccumulationMode.ACCUMULATING)
         # Extract and sum teamname/score pairs from the event data.
-        | 'ExtractAndSumScore' >> util.ExtractAndSumScore('team')
+        | 'ExtractAndSumScore' >> ExtractAndSumScore('team')
     )
 
 
@@ -152,7 +252,7 @@ class CalculateUserScores(beam.PTransform):
             trigger=trigger.Repeatedly(trigger.AfterCount(10)),
             accumulation_mode=trigger.AccumulationMode.ACCUMULATING)
         # Extract and sum username/score pairs from the event data.
-        | 'ExtractAndSumScore' >> util.ExtractAndSumScore('user')
+        | 'ExtractAndSumScore' >> ExtractAndSumScore('user')
     )
 
 
@@ -206,7 +306,7 @@ def run(argv=None):
     events = (
         p
         | 'ReadPubSub' >> beam.io.gcp.pubsub.ReadStringsFromPubSub(args.topic)
-        | 'ParseGameEventFn' >> beam.ParDo(util.ParseGameEventFn())
+        | 'ParseGameEventFn' >> beam.ParDo(ParseGameEventFn())
         | 'AddEventTimestamps' >> beam.Map(
             lambda elem: beam.window.TimestampedValue(elem, elem['timestamp']))
     )
@@ -221,9 +321,9 @@ def run(argv=None):
     (events  # pylint: disable=expression-not-assigned
      | 'CalculateTeamScores' >> CalculateTeamScores(
          args.team_window_duration, args.allowed_lateness)
-     | 'TeamScoresDict' >> beam.ParDo(util.TeamScoresDict())
+     | 'TeamScoresDict' >> beam.ParDo(TeamScoresDict())
      # Write the results to BigQuery.
-     | 'WriteTeamScoreSums' >> util.WriteToBigQuery(
+     | 'WriteTeamScoreSums' >> WriteToBigQuery(
          args.table_name + '_teams', args.dataset, teams_schema)
     )
 
@@ -237,7 +337,7 @@ def run(argv=None):
      | 'FormatUserScoreSums' >> beam.Map(
          lambda (user, score): {'user': user, 'total_score': score})
      # Write the results to BigQuery.
-     | 'WriteUserScoreSums' >> util.WriteToBigQuery(
+     | 'WriteUserScoreSums' >> WriteToBigQuery(
          args.table_name + '_users', args.dataset, users_schema)
     )
